+++ utils/isDebugDisable.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const code = fs.readFileSync(`${__dirname}/../src/ytAutoDark.js`);
+const code = fs.readFileSync(`${__dirname}/../src/core/logger.js`);
 
 if (code.includes('const debug = true;')) {
   // Failure